Update worker.go
I have changed the keyword VULNERABLE TO POSSIBLITY because it is more reasonable. Further if you store the output in some txt file and grep the key "POSSIBLITY" it become easier for users to filter out the results.

In case of VULNERABLE and NON VULNERABLE keyword you can not filter vulnerable results.